refactor: Rename components to multi-word
