+++ components/settings/settingsPanel.vue
@@ -13,9 +13,9 @@ import { useSettings } from '~~/stores/settings'
 import { NotificationPermission, useNotifications } from '~~/stores/notifications'
 import { useMobileSettings } from '~~/stores/platforms/mobileSettings'
 
-import ControlButton from '@/components/base/button.vue'
+import ControlButton from '~~/components/base/uiButton.vue'
 import SettingsItem from '~~/components/settings/settingsItem.vue'
-import Divider from '@/components/base/divider.vue'
+import Divider from '~~/components/base/uiDivider.vue'
 import { useEvents } from '~~/stores/events'
 import { useOpenPanels } from '~~/stores/openpanels'
 import { Control } from '~~/components/settings/types/settingsItem'
